Comments now are fetched in groups via blogs, blog title being the uid

+++ portfolio/src/main/java/com/google/sps/servlets/DeleteCommentServlet.java
@@ -31,7 +31,7 @@ public class DeleteCommentServlet extends HttpServlet {
 
   @Override
   public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {	
-    short id = Short.parseShort(request.getParameter("key"));
+    long id = Long.parseLong(request.getParameter("key"));
 
     Key commentEntityKey = KeyFactory.createKey("Comment", id);
     DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
